Changelog tool - extracted "transformCommit" function from "writer-options". [skip ci]
